requirements.txt
jax[cuda12]
jaxlib
flax
jraph
dask
trimesh
xarray
xarray-tensorstore
dm-tree
cartopy
chex
rtree
matplotlib
numpy
pandas
scipy
dinosaur-dycore
optax
grain
earthkit-data
earthkit-plots
cdsapi
cfgrib
eccodes
orbax-checkpoint
imageio